backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
sqlalchemy[asyncio]==2.0.35
asyncpg==0.29.0
pgvector==0.3.2
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.9
python-dotenv==1.0.1
groq==0.11.0
httpx==0.27.2
pypdf==4.3.1
python-docx==1.1.2
beautifulsoup4==4.12.3
requests==2.32.3
alembic==1.13.2
pydantic[email]==2.9.2
pydantic-settings==2.5.2
sentence-transformers==3.1.1
numpy==1.26.4
pillow==10.4.0
aiofiles==24.1.0
